Run static analysis using phpstan
Also fix some issues found in the first run

+++ src/Severity.php
@@ -34,7 +34,7 @@ class Severity {
             $severityLevels[] = self::$tags[strtolower($tag)] ?? self::WARNING;
         }
 
-        return min($severityLevels);
+        return (int) min($severityLevels);
     }
 
     /**
